Goのシリアライザは64bit環境前提で作っていました（符号有/無の32bit intをどちらも`int`で扱っていた）
ところが、dashboardで使っているGopherjsが32bit環境をエミュレートしているため、桁溢れするときに正しく扱えていませんでした。

32bit環境でも変わらないよう、Int/UIntはint64で扱うようにして、テストも32bitで通るようにしました。

+++ server/binary/event.go
@@ -139,7 +139,7 @@ func UnmarshalEvent(data []byte) (Event, int, error) {
 	if len(data) < 4 {
 		return nil, 0, xerrors.Errorf("data length not enough: %v", len(data))
 	}
-	seq := get32(data)
+	seq := int(get32(data))
 	data = data[4:]
 
 	return &RegularEvent{et, data}, seq, nil
@@ -194,7 +194,7 @@ func UnmarshalEvPeerReadyPayload(payload []byte) (int, error) {
 // - dict: last msg timestamps of each player.
 func NewEvPong(pingtime uint64, watchers uint32, lastMsg Dict) *SystemEvent {
 	payload := MarshalULong(pingtime)
-	payload = append(payload, MarshalUInt(int(watchers))...)
+	payload = append(payload, MarshalUInt(int64(watchers))...)
 	payload = append(payload, MarshalDict(lastMsg)...)
 
 	return &SystemEvent{
@@ -225,7 +225,7 @@ func UnmarshalEvPongPayload(payload []byte) (*EvPongPayload, error) {
 	if e != nil {
 		return nil, xerrors.Errorf("Invalid EvPong payload (watchers): %w", e)
 	}
-	pp.Watchers = uint32(d.(int))
+	pp.Watchers = uint32(d.(int64))
 	payload = payload[l:]
 
 	// lastmsg
